requirements.txt
aiosqlite==0.20.0
discord.py==2.4.0
jishaku==2.5.2
andreaw_utilities @ git+https://github.com/andrewthederp/Andreaws_utility_lib@777fda5709a5eb9f835ffba9f81a35ece9544bc8